# api/api_server/swagger.yaml
definitions:
...
  api.ApiOrder:
    properties:
      address:
        $ref: '#/definitions/api.ApiOrderAddress'
      promo_code:
        type: string
      recipient:
        $ref: '#/definitions/api.ApiOrderRecipient'
    type: object
  api.ApiOrderAddress:
    properties:
      address:
        type: string
    type: object
  api.ApiOrderRecipient:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
    type: object
  api.ApiPaginatorOrders:
    properties:
      count:
        type: integer
      page_num:
        type: integer
      sort_direction:
        type: string
      sort_key:
        type: string
    type: object
  api.ApiPaginatorProducts:
    properties:
      category:
        type: integer
      count:
        type: integer
      filter:
        $ref: '#/definitions/api.ApiProductFilter'
      page_num:
        type: integer
      sort_direction:
        type: string
      sort_key:
        type: string
    type: object
...
  api.ApiProductFilter:
    properties:
      is_discount:
        type: boolean
      is_new:
        type: boolean
      is_rating:
        type: boolean
      max_price:
        type: integer
      min_price:
        type: integer
    type: object
...
  api.ApiSearchQuery:
    properties:
      category:
        type: integer
      count:
        type: integer
      filter:
        $ref: '#/definitions/api.ApiProductFilter'
      page_num:
        type: integer
      query_string:
        type: string
      sort_direction:
        type: string
      sort_key:
        type: string
    type: object
...
  errors.Error:
    properties:
      error:
        type: string
    type: object
...
  usecase.CategoriesCatalog:
    properties:
      id:
        type: integer
      name:
        type: string
      next:
        items:
          $ref: '#/definitions/usecase.CategoriesCatalog'
        type: array
    type: object
...
  usecase.OrderAddress:
    properties:
      address:
        type: string
    type: object
  usecase.OrderNumber:
    properties:
      number:
        type: string
    type: object
  usecase.OrderRecipient:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
    type: object
...
  usecase.PlacedOrder:
    properties:
      address:
        $ref: '#/definitions/usecase.OrderAddress'
      date_added:
        type: string
      date_delivery:
        type: string
      id:
        type: integer
      order_number:
        $ref: '#/definitions/usecase.OrderNumber'
      product_images:
        items:
          $ref: '#/definitions/usecase.PreviewOrderedProducts'
        type: array
      status:
        type: string
      total_cost:
        type: integer
    type: object
...
  usecase.PreviewOrder:
    properties:
      address:
        $ref: '#/definitions/usecase.OrderAddress'
      price:
        $ref: '#/definitions/usecase.TotalPrice'
      products:
        items:
          $ref: '#/definitions/usecase.PreviewOrderedProducts'
        type: array
      recipient:
        $ref: '#/definitions/usecase.OrderRecipient'
    type: object
  usecase.PreviewOrderedProducts:
    properties:
      id:
        type: integer
      preview_image:
        type: string
    type: object
  usecase.Product:
    properties:
      category:
        type: integer
      category_path:
        items:
          $ref: '#/definitions/usecase.CategoriesCatalog'
        type: array
      count_reviews:
        type: integer
      description:
        type: string
      id:
        type: integer
      images:
        items:
          type: string
        type: array
      price:
        $ref: '#/definitions/usecase.ProductPrice'
      properties:
        type: string
      rating:
        type: number
      title:
        type: string
    type: object
  usecase.ProductPrice:
    properties:
      base_cost:
        type: integer
      discount:
        type: integer
      total_cost:
        type: integer
    type: object
...
  usecase.RangeOrders:
    properties:
      list_placed_orders:
        items:
          $ref: '#/definitions/usecase.PlacedOrder'
        type: array
      max_count_pages:
        type: integer
    type: object
  usecase.RangeProducts:
    properties:
      list_preview_products:
        items:
          $ref: '#/definitions/usecase.ViewProduct'
        type: array
      max_count_pages:
        type: integer
    type: object
...
  usecase.TotalPrice:
    properties:
      total_base_cost:
        type: integer
      total_cost:
        type: integer
      total_discount:
        type: integer
    type: object
...
  usecase.ViewProduct:
    properties:
      count_reviews:
        type: integer
      id:
        type: integer
      preview_image:
        type: string
      price:
        $ref: '#/definitions/usecase.ProductPrice'
      rating:
        type: number
      title:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /order:
    get:
      consumes:
      - application/json
      description: Получение заказа на основании корзины.
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Заказ успешно получен на основании корзины.
          schema:
            $ref: '#/definitions/usecase.PreviewOrder'
        "400":
          description: Некорректное тело запроса.
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: Непредвиденная ошибка сервера.
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - ApiKeyAuth: []
      summary: Получение заказа на основании корзины.
    post:
      consumes:
      - application/json
      description: Оформление пользовательского заказа.
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: Данные заказа.
        in: body
        name: Order
        required: true
        schema:
          $ref: '#/definitions/api.ApiOrder'
      produces:
      - application/json
      responses:
        "200":
          description: Заказ успешно оформлен.
          schema:
            $ref: '#/definitions/usecase.OrderNumber'
        "400":
          description: Некорректное тело запроса.
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: Непредвиденная ошибка сервера.
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - ApiKeyAuth: []
      summary: Оформление заказа.
  /product/{id}:
    get:
      consumes:
      - application/json
      description: Получение информации о товаре.
      parameters:
      - description: Id товара
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Информация о товаре успешно получена.
          schema:
            $ref: '#/definitions/usecase.Product'
        "400":
          description: Некорректное тело запроса.
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: Непредвиденная ошибка сервера.
          schema:
            $ref: '#/definitions/errors.Error'
      summary: Получение товара.
  /product/recommendations/{id}:
    post:
      consumes:
      - application/json
      description: Получение превью товаров.
      parameters:
      - description: Параметры пагинатора.
        in: body
        name: PaginatorProducts
        required: true
        schema:
          $ref: '#/definitions/api.ApiPaginatorProducts'
      produces:
      - application/json
      responses:
        "200":
          description: Превью товаров получены успешно.
          schema:
            $ref: '#/definitions/usecase.RangeProducts'
        "400":
          description: Некорректное тело запроса.
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: Непредвиденная ошибка сервера.
          schema:
            $ref: '#/definitions/errors.Error'
      summary: Получение превью товаров.
  /product/search:
    post:
      consumes:
      - application/json
      description: Поиск товаров.
      parameters:
      - description: Поисковый запрос.
        in: body
        name: SearchQuery
        required: true
        schema:
          $ref: '#/definitions/api.ApiSearchQuery'
      produces:
      - application/json
      responses:
        "200":
          description: Результат поиска.
          schema:
            $ref: '#/definitions/usecase.RangeProducts'
        "400":
          description: Некорректное тело запроса.
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: Непредвиденная ошибка сервера.
          schema:
            $ref: '#/definitions/errors.Error'
      summary: Поиск товаров.
...
  /user/order:
    post:
      consumes:
      - application/json
      description: Получение заказов пользователя.
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: Параметры пагинатора.
        in: body
        name: PaginatorOrders
        required: true
        schema:
          $ref: '#/definitions/api.ApiPaginatorOrders'
      produces:
      - application/json
      responses:
        "200":
          description: Список заказов данного пользователя успешно получен.
          schema:
            $ref: '#/definitions/usecase.RangeOrders'
        "400":
          description: Некорректное тело запроса.
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: Непредвиденная ошибка сервера.
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - ApiKeyAuth: []
      summary: Получение заказов.
...
swagger: "2.0"